--- sdl2-hyper-sonic-drivers/src/HyperSonicDrivers/audio/scummvm/SoundHandle.hpp
@@ -1,21 +1,0 @@
-#pragma once
-
-#include <cstdint>
-
-namespace HyperSonicDrivers::audio::scummvm
-{
-    /**
-     * A SoundHandle instance corresponds to a specific sound
-     * being played using the mixer. It can be used to control that
-     * sound (pause it, stop it, etc.).
-     * @see Mixer
-     */
-    class SoundHandle
-    {
-        friend class Channel;
-        friend class MixerImpl;
-        uint32_t _val;
-    public:
-        inline SoundHandle() : _val(0xFFFFFFFF) {}
-    };
-}
