+++ sdl2-hyper-sonic-drivers/src/HyperSonicDrivers/hardware/opl/mame/MameOPL3.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
-#include <HyperSonicDrivers/hardware/opl/scummvm/EmulatedOPL.hpp>
+#include <HyperSonicDrivers/hardware/opl/EmulatedOPL.hpp>
 #include <HyperSonicDrivers/hardware/opl/mame/ymfm/ymfm_opl.h>
 #include <HyperSonicDrivers/hardware/opl/mame/opl3/ymf262.h>
 
 namespace HyperSonicDrivers::hardware::opl::mame
 {
-    class MameOPL3 : public scummvm::EmulatedOPL
+    class MameOPL3 : public EmulatedOPL
     {
     private:
         // OPL3
